Locking is now local to the build method, and doesn't need cleaned up
Previous code had the chance to leave the locks active if the build
method died, and so had extra code to clean it up various places.

This version creates the lock as a local variable to the build method,
so it is guaranteed to get cleaned up when the method exits for any
reason.