Fixed file modes and path separator issues for Win32
For every slurp, spew, and open, I modified them to use the 'raw' binmode.

For the Directory asset, I updated multiple places where it was mixing
filenames and URI paths, which is fine on unix but on windows requires
a forward/backward slash conversion.

Also updated each testcase to remove the workdir before running the test
so that results of a previous run don't interfere with subsequent tests.